requirements.txt
absl-py==0.12.0
altair==4.1.0
argon2-cffi==20.1.0
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
idna==2.10
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.6.2
Pygments==2.9.0
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywinpty
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
Send2Trash==1.5.0
six==1.15.0
smmap==4.0.0
streamlit==0.82.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.5
uvicorn==0.14.0
validators==0.18.2
watchdog==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
